feat(core): 真交叉淡化 + 性能优化批次 + 栈溢出/竞态修复回归
修复：
- 自动跳曲递归改迭代（cherry-pick 987b6c7，该修复在另一分支未合入本分支）：
  play_entry ⇄ advance_queue 递归链改为 pick_next_entry + play_next_chain 迭代，
  修复连续坏轨栈溢出崩溃（RepeatOne+坏轨为无限递归必崩），带回 2 个回归测试
- playback_gen 防护统一进 stop_playback：seek/recover_output/stop_full/next_track
  四条路径此前未 bump 代际，旧消费者断链时误发 TrackChanged("") 导致误切歌/误报停止

性能优化：
- symphonia git main → crates.io 0.6.1（钉版可复现，去掉 [patch] 无双版本依赖）
- 解码热路径：rubato 重采样 f64→f32、mix_channels Cow 零拷贝、run_dsd 单次开文件
- probe_format_info 单次开文件同时取采样率+位深（bit-perfect 路径）
- 空闲省电：暂停时 tick 200ms→1s；DeviceMonitor 轮询 1.2s→5s
- EngineEvent::Spectrum Vec<f32>→[f32;16] 消除高频堆分配

新功能（真交叉淡化，默认关闭）：
- crossfade_ms>0 时，旧曲尾部与新曲头部在淡变窗口内按余弦曲线叠加混合
  （替代原仅新曲淡入）；引擎侧按位置武装/触发，消费者侧双源混合
- 完备回退：DoP/bit-perfect/流式/无预加载/新流失效 均降级为无间隙行为
- 测试：混合曲线余弦加权、无预加载回退单测 + 引擎双曲冒烟集成

验证：core clippy 0 / 307 测全过；desktop、mobile rust check 过

+++ core/src/dsp/speed.rs
@@ -11,10 +11,11 @@ use rubato::{InterpolationParameters, InterpolationType, Resampler, SincFixedOut
 pub struct SpeedChanger {
     speed: f32,
     channels: usize,
-    /// rubato 重采样器（speed 变化时重建）
-    resampler: Option<SincFixedOut<f64>>,
+    /// rubato 重采样器（speed 变化时重建）。用 f32：音频管线全程 f32，
+    /// 避免旧 f64 实现每帧的精度转换与双倍拷贝。
+    resampler: Option<SincFixedOut<f32>>,
     /// 每声道累积缓冲（等待足够帧数后送入 resampler）
-    accum: Vec<Vec<f64>>,
+    accum: Vec<Vec<f32>>,
     /// 输出缓冲（交错）
     output: Vec<f32>,
 }
@@ -69,7 +70,7 @@ impl SpeedChanger {
             oversampling_factor: 256,
             window: WindowFunction::BlackmanHarris2,
         };
-        self.resampler = Some(SincFixedOut::<f64>::new(ratio, params, 1024, channels));
+        self.resampler = Some(SincFixedOut::<f32>::new(ratio, params, 1024, channels));
         self.accum = vec![Vec::new(); channels];
     }
 
@@ -90,7 +91,7 @@ impl SpeedChanger {
         let frames = input.len() / channels;
         for f in 0..frames {
             for c in 0..channels {
-                self.accum[c].push(input[f * channels + c] as f64);
+                self.accum[c].push(input[f * channels + c]);
             }
         }
 
@@ -100,7 +101,7 @@ impl SpeedChanger {
             if self.accum[0].len() < needed {
                 break;
             }
-            let waves_in: Vec<Vec<f64>> = self
+            let waves_in: Vec<Vec<f32>> = self
                 .accum
                 .iter_mut()
                 .map(|buf| buf.drain(..needed).collect())
@@ -111,7 +112,7 @@ impl SpeedChanger {
                     self.output.reserve(out_frames * channels);
                     for f in 0..out_frames {
                         for c in 0..channels {
-                            self.output.push(waves_out[c][f] as f32);
+                            self.output.push(waves_out[c][f]);
                         }
                     }
                 }
